refactor: extract OP transaction pool types to separate crate
