+++ crates/optimism/node/Cargo.toml
@@ -40,6 +40,7 @@ reth-tasks = { workspace = true, optional = true }
 reth-optimism-payload-builder.workspace = true
 reth-optimism-evm.workspace = true
 reth-optimism-rpc.workspace = true
+reth-optimism-txpool.workspace = true
 reth-optimism-chainspec.workspace = true
 reth-optimism-consensus.workspace = true
 reth-optimism-forks.workspace = true
@@ -92,16 +93,17 @@ futures.workspace = true
 [features]
 default = ["reth-codec"]
 optimism = [
-    "reth-optimism-evm/optimism",
-    "reth-optimism-payload-builder/optimism",
-    "revm/optimism",
-    "reth-optimism-rpc/optimism",
-    "reth-engine-local/op",
-    "reth-optimism-consensus/optimism",
-    "reth-db/optimism",
-    "reth-optimism-node/optimism",
-    "reth-node-core/optimism",
-    "reth-optimism-primitives/optimism",
+	"reth-optimism-evm/optimism",
+	"reth-optimism-payload-builder/optimism",
+	"revm/optimism",
+	"reth-optimism-rpc/optimism",
+	"reth-engine-local/op",
+	"reth-optimism-consensus/optimism",
+	"reth-db/optimism",
+	"reth-optimism-node/optimism",
+	"reth-node-core/optimism",
+	"reth-optimism-primitives/optimism",
+	"reth-optimism-txpool/optimism"
 ]
 asm-keccak = [
     "reth-primitives/asm-keccak",
