+++ crates/optimism/node/src/lib.rs
@@ -23,7 +23,7 @@ pub use engine::OpEngineTypes;
 pub mod node;
 pub use node::{OpNetworkPrimitives, OpNode};
 
-pub mod txpool;
+pub use reth_optimism_txpool as txpool;
 
 /// Helpers for running test node instances.
 #[cfg(feature = "test-utils")]
