+++ crates/optimism/txpool/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+name = "reth-optimism-txpool"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description = "OP-Reth Transaction Pool"
+
+[lints]
+workspace = true
+
+[dependencies]
+# ethereum
+alloy-consensus.workspace = true
+alloy-eips.workspace = true
+alloy-primitives.workspace = true
+alloy-rpc-types-eth.workspace = true
+
+# reth
+reth-chainspec.workspace = true
+reth-primitives-traits.workspace = true
+reth-storage-api.workspace = true
+reth-transaction-pool.workspace = true
+
+# revm
+revm.workspace = true
+
+# optimism
+op-alloy-consensus.workspace = true
+op-alloy-flz.workspace = true
+reth-optimism-evm.workspace = true
+reth-optimism-forks.workspace = true
+reth-optimism-primitives = { workspace = true, features = ["reth-codec"] }
+
+# misc
+c-kzg.workspace = true
+derive_more.workspace = true
+parking_lot.workspace = true
+
+[features]
+optimism = [
+	"reth-optimism-evm/optimism",
+	"reth-optimism-primitives/optimism",
+	"revm/optimism"
+]
